Module Current_github.Api

type t

Configuration for accessing GitHub.

module Status : sig ... end

module Commit : sig ... end

module Repo : sig ... end

val of_oauth : string -> t

of_oauth token is a configuration that authenticates to GitHub using token.

val exec_graphql : ?⁠variables:(string * Yojson.Safe.t) list -> t -> string -> Yojson.Safe.t Lwt.t

exec_graphql t query executes query on GitHub.

val head_commit : t -> Repo_id.t -> Commit.t Current.t

head_commit t repo evaluates to the commit at the head of the default branch in repo.

val ci_refs : t -> Repo_id.t -> Commit.t list Current.t

ci_refs t repo evaluates to the list of branches and open PRs in repo.

val cmdliner : t Cmdliner.Term.t

Command-line options to generate a GitHub configuration.