Module Api.Anonymous

Perform Anonymous request to GitHub.

val head_of : Repo_id.t -> Ref.t -> Current_git.Commit_id.t Current.t

head_of repo ref is the head commit of repo/ref. No API token is used to access this, so it only works for public repositories. You are responsible for adding a web-hook so that input_webhook gets called whenever the commit changes.