Module Api.Commit

type t

val id : t -> Current_git.Commit_id.t

The commit ID, which can be used to fetch it.

val set_status : t Current.t -> string -> Status.t Current.t -> unit Current.t

set_status commit context status sets the status of commit/context to status.

val owner_name : t -> string

owner_name t is the "owner/name" string identifying the repository.

val repo_id : t -> Repo_id.t

Like owner_name, but as a Repo_id.t.

val hash : t -> string

hash t is the Git commit hash of t.

val pp : t Fmt.t

val compare : t -> t -> int

val uri : t -> Uri.t

uri t is a URI for the GitHub web page showing t.