Module Api.Monitor

Monitor a GraphQL query for changes on webhooks.

Parameters

module Query : GRAPHQL_QUERY

Signature

val get : t -> Repo_id.t -> Query.result Current.Primitive.t

get t repo is the primitive for observing the result of the GraphQL Query.query. The result is cached (so calling it twice will return the same primitive), with the same lifetime as t.