Module Api.Repo

type nonrec t = t * Repo_id.t

val id : t -> Repo_id.t

val pp : t Fmt.t

val compare : t -> t -> int

val ci_refs : ?staleness:Duration.t -> t Current.t -> Commit.t list Current.t

ci_refs t evaluates to the list of branches and open PRs in t, excluding gh-pages.

parameter staleness
If given, commits older than this are excluded. Note: the main branch commit is always included, even if stale.

val head_commit : t Current.t -> Commit.t Current.t

head_commit t evaluates to the commit at the head of the default branch in t.