Module type Api.GRAPHQL_QUERY

A GraphQL query to be monitored for changes

type result

The result type produced by the query

val name : string

A short name describing the query (for logging)

val query : string

The GraphQL query that will be monitored. The variable $owner and $name are available and bound to the repository's owner and name.

Furthermore, this query will be wrapped inside a template to also report the rate limitations:

query($owner: String!, $name: String!) {
  rateLimit { ... }
  <<query>>
}

val of_yojson : t -> Repo_id.t -> Yojson.Safe.t -> result

of_yojson t repo json parses the json into a result.