Module Current_github__Api

module Status : sig ... end

module Commit : sig ... end

module Ref : sig ... end

module Ref_map : Stdlib.Map.S with type Ref_map.key = Ref.t

type t

val of_oauth : string -> t

val exec_graphql : ?⁠variables:(string * Yojson.Safe.t) list -> t -> string -> Yojson.Safe.t Lwt.t

val head_commit : t -> Current_github.Repo_id.t -> Commit.t Current.t

val refs : t -> Current_github.Repo_id.t -> Commit.t Ref_map.t Current.Primitive.t

val head_of : t -> Current_github.Repo_id.t -> [ `Ref of string | `PR of int ] -> Commit.t Current.t

val ci_refs : t -> Current_github.Repo_id.t -> Commit.t list Current.t

val cmdliner : t Cmdliner.Term.t

module Repo : sig ... end

module Anonymous : sig ... end

val read_file : string -> string

read_file path is the contents of the file at path (just a utility function; should be moved elsewhere).

type token = {

token : (string, [ `Msg of string ]) Stdlib.result;
A token to include in the "Authorization" header, or an error if we failed to get a token.

expiry : float option;
token is valid until this time. If token is an Error then this is the earliest time to try again. If None, token does not expire.

}

val get_token : t -> (string, [ `Msg of string ]) Stdlib.result Lwt.t

get_token t returns the cached token for t, or fetches a new one if it has expired.

val input_webhook : Yojson.Safe.t -> unit

Call this when we get a "pull_request", "push" or "create" webhook event.

val v : get_token:(unit -> token Lwt.t) -> string -> t

v ~get_token account is a configuration that uses get_token when it needs to get or refresh the API token. Note: get_token can return a failed token, in which case the expiry time says when to try again. If get_token instead raises an exception, this is turned into an error token with a 1 minute expiry.

parameter account

This is a string used to label point counters in Prometheus.