Module Api.Commit

type t

val id : t -> Current_git.Commit_id.t

The commit ID, which can be used to fetch it.

val set_status : t Current.t -> string -> Status.t Current.t -> unit Current.t

set_status commit context status sets the status of commit/context to status.

val owner_name : t -> string

owner_name t is the "owner/name" string identifying the repository.

val repo_id : t -> Repo_id.t

Like owner_name, but as a Repo_id.t.

val hash : t -> string

hash t is the Git commit hash of t.

val committed_date : t -> string

committed_date t is the datetime when t was committed

val message : t -> string

message t is the Git commit message of t.

val pp : t Fmt.t

pp t pretty print the commit as "owner/name id hash committed_date (Commit message)"

val pp_short : t Fmt.t

pp_short t pretty print the commit as "owner/name hash"

val compare : t -> t -> int

val uri : t -> Uri.t

uri t is a URI for the GitHub web page showing t.

val pr_name : t -> string option

pr_name t is the name of the ref that the commit belongs to if it is a PR, and None if it is a branch

val branch_name : t -> string option

branch_name t is the name of the ref that the commit belongs to if it is a branch, and None if it is a PR