Module Current_github.App

A GitHub Application.

type t

Configuration for a GitHub application.

val webhook_secret : t -> string

Webhook secret to validate payloads from GitHub.

val cmdliner : t Cmdliner.Term.t

Command-line options to generate a GitHub app configuration.

val cmdliner_opt : t option Cmdliner.Term.t

Like cmdliner, but the arguments are all optional.

val installation : t -> account:string -> int -> Installation.t

installation t ~account id gives access to the API for installation id. Note: this generates a fresh value on each call and should not be used inside a pipeline. It is intended to be called once from your main function, for apps with only a single installation.

parameter account
The GitHub account that installed the application.

val installations : t -> Installation.t list Current.t

installations t evaluates to the list of installations for this app.